+++ src/modules/gecko-gui/src/main/java/gecko/geckocircuits/datacontainer/DataContainerSimple.java
@@ -113,7 +113,7 @@ public class DataContainerSimple extends AbstractDataContainer implements DataCo
         HiLoData hiLoData = null;
 
         for (int index = columnStart; index < columnStop; index++) {
-            hiLoData = HiLoData.mergeFromValue(hiLoData, (float) _data[row][index]);
+            hiLoData = HiLoData.mergeFromValue(hiLoData, _data[row][index]);
         }
 
         return hiLoData;
