+++ src/modules/gecko-simulation-core/src/main/java/gecko/core/datacontainer/DataContainerSimple.java
@@ -80,7 +80,7 @@ public class DataContainerSimple extends AbstractDataContainer implements DataCo
         assert columnStart <= columnStop;
         HiLoData hiLoData = null;
         for (int index = columnStart; index < columnStop; index++) {
-            hiLoData = HiLoData.mergeFromValue(hiLoData, (float) _data[row][index]);
+            hiLoData = HiLoData.mergeFromValue(hiLoData, _data[row][index]);
         }
         return hiLoData;
     }
